[ok] dr.monero - histogram avg and doc

+++ utils/dr-monero/utils.cpp
@@ -42,9 +42,9 @@ vector<long double> utils::prepareHistogramData(vector<long double> t, vector<lo
     frame--; //TODO frame!=1
     double value=0;
 
-    size_t size = t.size()-1;
+    int size = t.size()-1;
 
-    for (size_t i=0; i<size; i++) {
+    for (int i=0; i<size; i++) {
         value+=b.at(i);
 
         if( int(t.at(i+1)) - int(t.at(i))  == 1  )  {
